+++ pkg/platform/helpers.go
@@ -4,35 +4,22 @@
 package platform
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"os"
-	"os/exec"
-	"strings"
 	"time"
 )
 
-// commandTimeout bounds every shell-out so a wedged command — a stuck mirror, a tool that prompts on /dev/tty — can
-// never hang the run indefinitely. It is a generous backstop, not a per-operation deadline (large installs are
-// legitimate). TODO: make it configurable via the run environment.
-const commandTimeout = 30 * time.Minute
-
-// confirmReplies is how many "y" answers are piped to a command's stdin — a backstop that auto-confirms any prompt a
-// tool raises despite its non-interactive flags, so an op the plan requested proceeds instead of aborting on EOF.
-// Far more than any command asks; the stream ends in EOF afterward. apt's dangerous conffile prompt is handled
-// separately and safely by DEBIAN_FRONTEND (keep-old default), not by this stream.
-const confirmReplies = 4096
-
-// refreshTTL is the staleness threshold for the automatic index refresh: a leaf whose local index is older than this
-// refreshes before an index-consuming operation. A single default knob for now; promote to a per-leaf value if
-// managers need to diverge.
+// refreshTTL is the staleness threshold for the automatic index refresh.
+//
+// A leaf whose local index is older than this refreshes before an index-consuming operation. A single default knob
+// for now; promote to a per-leaf value if managers need to diverge.
 const refreshTTL = 24 * time.Hour
 
-// unknownIndexAge is the age reported for an index that cannot be stat'd (never built, or an unreadable path): well
-// past [refreshTTL], so the gate treats it as stale and refreshes.
+// unknownIndexAge is the age reported for an index that cannot be stat'd (never built, or an unreadable path).
 //
-//nolint:unused // read by the Linux-tagged leaves (apt, pacman) in linux_managers_linux.go; invisible to the darwin analysis.
+// Well past [refreshTTL], so the gate treats it as stale and refreshes.
+//
+//nolint:unused // read by the Linux-tagged leaves in linux_managers_linux.go; invisible to non-Linux analyses.
 const unknownIndexAge = 365 * 24 * time.Hour
 
 // indexAgeOf returns how long ago `path` was last modified, or [unknownIndexAge] when it cannot be stat'd.
@@ -46,7 +33,7 @@ const unknownIndexAge = 365 * 24 * time.Hour
 // Returns:
 //   - `time.Duration`: the age since last modification, or [unknownIndexAge] when the path is unreadable.
 //
-//nolint:unused // called by the Linux-tagged leaves (apt, pacman) in linux_managers_linux.go; invisible to the darwin analysis.
+//nolint:unused // called by the Linux-tagged leaves in linux_managers_linux.go; invisible to non-Linux analyses.
 func indexAgeOf(path string) time.Duration {
 
 	info, err := os.Stat(path)
@@ -55,63 +42,6 @@ func indexAgeOf(path string) time.Duration {
 	}
 
 	return time.Since(info.ModTime())
-}
-
-// runShellCommand executes a shell command via bash, optionally with sudo, capturing the result.
-//
-// It captures stdout, stderr, and the exit code into a [Result]. Used by every Linux/Darwin
-// [PackageManager] and [ServiceManager] mutator. The command string is passed to `bash -c` directly; callers are
-// responsible for safe quoting.
-//
-// Hang safety, in layers: the call is bounded by [commandTimeout]; an endless `y\n` stream on stdin auto-confirms
-// any prompt a tool raises despite its non-interactive flags — a backstop for the ones port's `-N` misses — so the
-// op proceeds rather than blocking or aborting on EOF; sudo runs with `-n`, so it fails fast instead of prompting
-// for a password on the tty (credential handling is the elevation model's job — see TODO); and
-// `DEBIAN_FRONTEND=noninteractive` keeps apt quiet (and safely keeps modified config files). Callers still pass
-// per-tool non-interactive flags (apt `-y`, pacman `--noconfirm`, port `-N`).
-//
-// It is a package var, not a plain func, so tests can substitute a recording fake.
-var runShellCommand = func(command string, sudo bool) Result {
-
-	ctx, cancel := context.WithTimeout(context.Background(), commandTimeout)
-	defer cancel()
-
-	var cmd *exec.Cmd
-	if sudo {
-		// TODO(elevation): centralize this. Route privileged execution through the ElevationOffer/Elevator — one
-		// credential-cached, policy-governed, audited sudo session — instead of every command inlining `sudo -n`.
-		cmd = exec.CommandContext(ctx, "sudo", "-n", "bash", "-c", command) //nolint:gosec // G204: shell command from internal caller
-	} else {
-		cmd = exec.CommandContext(ctx, "bash", "-c", command) //nolint:gosec // G204: shell command from internal caller
-	}
-
-	cmd.Env = append(cmd.Environ(), "DEBIAN_FRONTEND=noninteractive")
-	cmd.Stdin = strings.NewReader(strings.Repeat("y\n", confirmReplies))
-
-	var stdout, stderr strings.Builder
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-
-	err := cmd.Run()
-	code := 0
-	if err != nil {
-		if exitErr, ok := errors.AsType[*exec.ExitError](err); ok {
-			code = exitErr.ExitCode()
-		} else {
-			code = -1
-		}
-	}
-
-	if ctx.Err() == context.DeadlineExceeded {
-		fmt.Fprintf(&stderr, "\ncommand timed out after %s", commandTimeout)
-	}
-
-	return Result{
-		OK:     code == 0,
-		Stdout: strings.TrimSuffix(stdout.String(), "\n"),
-		Stderr: strings.TrimSuffix(stderr.String(), "\n"),
-		Code:   code,
-	}
 }
 
 // bracket runs a best-effort batch package operation and returns one [Receipt] per package.
@@ -153,7 +83,8 @@ func bracket(packages []PURL, token func(p PURL) string, version func(name strin
 
 		var err error
 		if !satisfied(post) {
-			err = fmt.Errorf("platform: %s/%s did not reach the requested state (post=%q): %s", p.Type, p.Name, post, result.Stderr)
+			err = fmt.Errorf("platform: %s/%s did not reach the requested state (post=%q): %s",
+				p.Type, p.Name, post, result.Stderr)
 		}
 
 		receipts[i] = Receipt{Purl: p, PriorVersion: prior[i], Version: post, Err: err}
